refactor(broker): extract budget-candidate-index.ts to stay under 1500 LOC
Combining PR #845 (bigint accumulators + decimal-string wire shape) and
PR #846 (scope-keyed budget candidate indexes) into one branch pushed
`replay-check.ts` from 1395 (pre-845) and 1417 (pre-846) to 1586 LOC,
86 above the 1500-LOC ceiling enforced by scripts/check-file-size.sh.
The allowlist is forward-only and currently empty; adding a first
entry would be a CONTRIBUTING.md violation.

Extract the budget-candidate-index module: BudgetCandidateIndexes type
+ createBudgetCandidateIndexes, addBudgetToIndex, removeBudgetFromIndex,
replaceBudgetInIndex, __createBudgetCandidateIndexesForTesting, plus
the two private subject-index helpers. The new module imports
`type ReplayedBudget` from replay-check.ts (type-only, no runtime cycle)
and replay-check.ts imports the runtime helpers + re-exports the test
seam so existing test imports keep working unchanged.

Net: replay-check.ts 1497 LOC (under 1500); budget-candidate-index.ts
123 LOC. All 248 broker tests green, typecheck + biome clean, file-size
hook passes.
